Develop

+++ Diagram.SRC/Diagram/Src/Forms/LineWidthForm.cs
@@ -1,16 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Forms;
-
-/*
- 
-*/
 
 namespace Diagram
 {
@@ -73,9 +62,9 @@ namespace Diagram
                 this.trackbarStateChanged(this.trackBar1.Value);
         }
 
-        public void setValue(int value)
+        public void setValue(long value)
         {
-            this.trackBar1.Value = value;
+            this.trackBar1.Value = (int)value; // TODO: scale long to int
         }
         
     }
